fix: prevent errors with row-specific ghost in combination with grouping, selection and trees (#241)
This is just preventing runtime errors and not supporting the combination.
Since we use `undefined` to mark loading rows, we do not have access to group keys and alike.
So it is not possible yet to have loading rows inside a group or a tree structure.

To solve this, we need to enhance the API.