refactor: base internal row length on rowCount property (#261)
The number of potentially rendered rows should be based on the `rowCount`
property which can be supplied externally for server paging.
This is needed for ghost loading as apps may provide rows
which are shorter than the `rowCount` and thus than the rendered rows.
By filling up the internal rows with `undefined` the ghost loading
works even though the rows provided are too short.