feat: export pager component for reuse in custom footer
When building a custom footer, our pagination is no longer included,
but typically still required.
With previous versions, apps could re-use our internal pager,
as just everything was exported. This was stopped for good reasons.

This commit now adjusts the pager for being exported again.
Now the pager automatically picks the table state,
leading to the removal of all input/outputs.
The selector is also changed to `ngx-datatable-pager` to match
the pattern of the already exported components.

Closes #324